==> marine/tuna/tuna_farms/STEP1_fao_tuna_production.Rmd <==
---
title: "FAO marine tuna production"
author: "Gage Clawson (NCEAS, UCSB, OHI)"
date: "12/11/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

All data is ultimately adjusted to FAO production data.  This script cleans the fao mariculture data and subsets it to to the relevant years of data and species (tuna).

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}

library(tidyverse)
library(data.table)
library(countrycode)
library(here)
library(stringr)

path <- "/home/shares/food-systems/Food_footprint/_raw_data"
source(here('marine/salmon/salmon_farm/R/fao_mar_clean.R')) # functions specific to mariculture dealing with compound countries
source('https://raw.githubusercontent.com/OHI-Science/ohiprep_v2019/gh-pages/workflow/R/fao_fxn.R') # function for cleaning FAO files
source(here('workflow/common.R'))

```

Subset the data to include only the species (tuna) and countries we are interested in. Then summarize by total tonnes of tuna per country/year. (There are multiple tuna species)

```{r}

fao <- read.csv(here("data/fao_aquaculture_clean.csv")) %>%
  filter(environment %in% c("Marine", "Brackishwater"))


## Read in fao data and get a species list for tuna
species_list <- read.csv(here("marine/STEP1_species_groups/int/marine_groups.csv"))  %>%
  distinct(species, aq_group)

# the list of species we are including
tuna_list <- species_list$species[species_list$aq_group=="tuna"]
tuna_list

## write out tuna taxa and production
fao_tuna_taxa_prod <- fao %>%
  filter(species %in% tuna_list) %>%  
  filter(!is.na(value)) %>%
  arrange(-value, country) %>%
  filter(value>=500) %>%  # countries with less do not appear to be stable producers at this point
  group_by(country, iso3c, species, year) %>%
  dplyr::filter(year == 2017) %>%
  summarize(fao_tonnes_production = sum(value)) %>%
  arrange(country, iso3c, species, year)
  
write.csv(fao_tuna_taxa_prod, here("marine/tuna/tuna_farms/data/fao_mariculture_tuna_taxa_production.csv"), row.names=FALSE)

# include only tuna
fao_tuna <- fao %>%
  filter(species %in% tuna_list) %>%  
  filter(!is.na(value)) %>%
  arrange(-value, country) %>%
  filter(value>=500) %>%  # countries with less do not appear to be stable producers at this point
  group_by(country, iso3c, year) %>%
  summarize(fao_tonnes_production = sum(value)) %>%
  arrange(country, iso3c, year) 
 
write.csv(fao_tuna, here("marine/tuna/tuna_farms/data/fao_mariculture_tuna.csv"), row.names=FALSE)
        
 explore <- read_csv(here("marine/tuna/tuna_farms/data/fao_mariculture_tuna.csv")) %>%
   filter(year==2017) %>%
   arrange(-fao_tonnes_production) %>%
   mutate(percent = round((fao_tonnes_production/sum(fao_tonnes_production))*100, 2))
 sum(explore$fao_tonnes_production) # 37072.56
 unique(explore$country) # 7 countries 
```

```{r}
## explore out bivalve taxa and production 
fao_tuna_taxa_prod <- fao %>%
  filter(species %in% tuna_list) %>%  
  filter(!is.na(value)) %>%
  arrange(-value, country) %>%
  group_by(country, year) %>%
  summarise(value = sum(value)) %>%
  filter(value>=500) %>% # countries with less do not appear to be stable producers at this point
  filter(year == 2017)

sum(fao_tuna_taxa_prod$value) # 37072.56

unique(fao_tuna_taxa_prod$country) # 16 countries 
```
