==> marine/tuna/tuna_farms/STEP3_gapfill_tuna.Rmd <==
---
title: "Gapfilling the number of farms to add for tuna countries based on tonnes of production"
author: "Gage Clawson (NCEAS, UCSB, OHI)"
date: "12/14/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---


This script takes the number of farms information collected in STEP2 and calculates a global tonnes of production per farm, and calculates the number of remaining farms needed to place in our placement model. 


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Read in all the relevant farm datasets 

```{r}

farm_files <- list.files(here("marine/tuna/tuna_farms/data/known_farm_sites_2_4_5"), full=TRUE)
all_known_farms = plyr::ldply(farm_files, read_csv) %>%
  mutate(type = "tuna")

number_files <- list.files(here("marine/tuna/tuna_farms/data/known_number_farms_3"), full=TRUE)
all_number_3 = plyr::ldply(number_files, read_csv)

number_files_2.5 <- list.files(here("marine/tuna/tuna_farms/data/known_subset_2.5"), full=TRUE)
all_number_2.5 = plyr::ldply(number_files_2.5, read_csv)

subset_farms_files <- list.files(here("marine/tuna/tuna_farms/data/subset_known_add_points_4"), full=TRUE)
subset_add_4 = plyr::ldply(subset_farms_files, read_csv)

no_info_files <- list.files(here("marine/tuna/tuna_farms/data/no_info_1"), full=TRUE)
no_info_1 = plyr::ldply(no_info_files, read_csv)

subset_gapfill_2 <- all_known_farms %>%
  filter(data_quality == 2)
```

Read in relevant production data
```{r}
tuna_tonnes <- read_csv(here::here("marine/tuna/tuna_farms/data/fao_mariculture_tuna.csv")) %>%
  mutate(type = "tuna") %>%
  filter(year == 2017)
```

Calculate a global tonnes per farm for tuna

```{r}
## Filter for data_quality = 4 or 5 and group by country to get number of farms
comprehensive_farms <- all_known_farms %>%
  filter(data_quality %in% c(4,5)) %>% 
  group_by(iso3c, type, data_quality) %>%
  summarise(n_farms = n()) %>% ## now we can see the places that has additional farms to add
  ungroup()

## Join this with the additional farms we need to add to some places, so we can calculate a comprehensive tonnes/farm for tuna 

data_quality_3 <- all_number_3 %>%
    rename("n_farms" = "num_additional_farms")

data_quality_4 <- subset_add_4 %>%
    rename("n_farms" = "num_additional_farms")

comprehensive_farms_total <- comprehensive_farms %>%
  full_join(data_quality_3) %>%
  full_join(data_quality_4) %>%
  group_by(iso3c, type) %>%
  summarise(n_farms = sum(n_farms)) %>%
  ungroup()
  

tonnes_per_farm <- comprehensive_farms_total %>%
  left_join(tuna_tonnes) %>%
  mutate(tonnes_per_farm = fao_tonnes_production/n_farms)

write.csv(tonnes_per_farm, "output/tuna_comp_tonnes_per_farm.csv", row.names = FALSE)


avg_global_tuna_tonnes_per_farm = mean(tonnes_per_farm$tonnes_per_farm)
avg_global_tuna_tonnes_per_farm # 277.4158

med_global_tuna_tonnes_per_farm = median(tonnes_per_farm$tonnes_per_farm)
med_global_tuna_tonnes_per_farm # 172.3696
```

Now let's gapfill those that are data quality = 2 and data quality = 1, and data quality = 2.5 (if applicable)

```{r}
## start with the subset of locations that need gapfilling based on tonnes/farm
# there are none for tuna 

## now lets do the locations that have no info and need gapfilling based on tonnes/farm
# there are none for tuna 

## Wrangle those that just need a specific number of points added (data_quality = 4)

number_add_points_4 <- subset_add_4 %>%
  dplyr::select(-data_quality, -total_farms) %>%
  mutate(estimated = 0, 
         avg_tonnes_per_farm = avg_global_tuna_tonnes_per_farm)

```

Now join the gapfilled number of farms with data quality = 3 and write the final file of points that need to allocated 

```{r}
number_additional_points <- all_number_3 %>%
  mutate(estimated = 0,
         avg_tonnes_per_farm = avg_global_tuna_tonnes_per_farm) %>%
  dplyr::select(-data_quality) %>%
  rbind(number_add_points_4)
  

## write this out to our data folder

write.csv(number_additional_points, here("marine/tuna/tuna_farms/output/tuna_points_to_add_final.csv"), row.names = FALSE)
```

Now lets make the file with all of the known farms 

```{r}
all_known_farms_final <- all_known_farms %>%
  dplyr::select(-data_quality)


write.csv(all_known_farms_final, here("marine/tuna/tuna_farms/output/all_known_tuna_farms_final.csv"), row.names = FALSE)

```



Data check 

```{r}
test <- all_known_farms %>%
  group_by(iso3c) %>%
  summarise(n_farms = n()) %>%
  dplyr::select(iso3c)


test_2 <- number_additional_points %>%
  dplyr::select(iso3c)


test_all <- rbind(test, test_2)
length(unique(test_all$iso3c)) ## 7 countries


length(unique(tuna_tonnes$iso3c)) ## 7 countries


## Perfect

```
